AILogDemo/requirements.txt
kafka-python==2.0.2
scikit-learn==1.5.0
pandas==2.2.2
requests==2.32.3
numpy==1.26.4
